+++ p2panda-rs/src/schema/mod.rs
@@ -9,15 +9,16 @@
 #[cfg(not(target_arch = "wasm32"))]
 use cddl::validator::cbor;
 
+#[allow(clippy::module_inception)]
+mod cddl_builder;
 mod error;
 mod operation;
-#[allow(clippy::module_inception)]
-mod schema;
 mod schema_id;
+mod system_schema;
 
-pub use error::SchemaError;
+pub use cddl_builder::CDDLBuilder;
+pub use error::{SchemaValidationError, SystemSchemaError};
 pub use operation::OPERATION_SCHEMA;
-pub use schema::{Schema, SchemaBuilder, Type, ValidateOperation};
 pub use schema_id::SchemaId;
 
 /// Checks CBOR bytes against CDDL schemas.
@@ -25,7 +26,7 @@ pub use schema_id::SchemaId;
 /// This helper method also converts validation errors coming from the `cddl` crate into an
 /// concatenated error operation and returns it.
 #[cfg(not(target_arch = "wasm32"))]
-pub fn validate_schema(cddl_schema: &str, bytes: Vec<u8>) -> Result<(), SchemaError> {
+pub fn validate_schema(cddl_schema: &str, bytes: Vec<u8>) -> Result<(), SchemaValidationError> {
     match cddl::validate_cbor_from_slice(cddl_schema, &bytes) {
         Err(cbor::Error::Validation(err)) => {
             let err_str = err
@@ -36,13 +37,13 @@ pub fn validate_schema(cddl_schema: &str, bytes: Vec<u8>) -> Result<(), SchemaEr
                         // format macro.
                         //
                         // See: https://github.com/anweiss/cddl/blob/main/src/validator/cbor.rs#L100
-                        .replace("\"", "'")
+                        .replace('"', "'")
                 })
                 .collect::<Vec<String>>();
 
-            Err(error::SchemaError::InvalidSchema(err_str))
+            Err(SchemaValidationError::InvalidSchema(err_str))
         }
-        Err(cbor::Error::CBORParsing(_err)) => Err(error::SchemaError::InvalidCBOR),
+        Err(cbor::Error::CBORParsing(_err)) => Err(SchemaValidationError::InvalidCBOR),
         Err(cbor::Error::CDDLParsing(err)) => {
             panic!("Parsing CDDL error: {}", err);
         }